+++ migrations/019_mirror_host.sql
@@ -1,0 +1,2 @@
+ALTER TABLE download_file_mirrors ADD COLUMN host VARCHAR(255) AFTER file_id;
+